+++ packages/core/src/phone-field/PhoneFieldControlState.ts
@@ -1,0 +1,8 @@
+import PhoneFieldCountry from './PhoneFieldCountry'
+
+export default interface PhoneFieldControlState {
+  selectedCountry: PhoneFieldCountry | null
+  focusedCountry: PhoneFieldCountry | null
+  showCountries: boolean
+  focused: boolean
+}