+++ packages/desktop/src/pagination/PaginationLink.tsx
@@ -1,6 +1,6 @@
 import React, {FC} from 'react'
 
-import {LinkControl, Lnk, Card, Value, RenderChild, Flex} from '@qiwi/pijma-core'
+import {styled, LinkControl, Lnk, Card, Value, RenderChild, Flex, Stub, CardOptions} from '@qiwi/pijma-core'
 
 export interface PaginationLinkProps {
   page: number
@@ -8,6 +8,7 @@ export interface PaginationLinkProps {
   width?: Value
   href?: string
   shadowed?: boolean
+  stub?: boolean
   onClick?: (href?: string, target?: string, download?: string | boolean, rel?: string) => void
   children: RenderChild<{
     disabled: boolean
@@ -16,44 +17,53 @@ export interface PaginationLinkProps {
   }>
 }
 
-const CardLink = Card.withComponent(Lnk)
+const CardLink = styled(Card, CardOptions)().withComponent(Lnk)
 
 export const PaginationLink: FC<PaginationLinkProps> = props => (
   <LinkControl
-    href={props.href}
-    onClick={props.onClick}
+    href={props.stub ? undefined : props.href}
+    onClick={props.stub ? undefined : props.onClick}
     children={renderProps => (
       <CardLink
+        as={props.stub ? 'div' : 'a'}
         transition="all 100ms cubic-bezier(0.4, 0.0, 0.2, 1)"
         height={12}
         width={props.width}
         display="inline-flex"
-        cursor={props.disabled ? 'default' : 'pointer'}
+        cursor={props.stub || props.disabled ? 'default' : 'pointer'}
         bg={(renderProps.hover || renderProps.focus) && !props.disabled ? '#f5f5f5' : undefined}
         s={
           (renderProps.hover || renderProps.focus) && !props.disabled
             ? props.shadowed ? 'inset 0 1px 0 0 #e6e6e6, 0 -1px 0 1px #e6e6e6' : '0px -1px 0 1px #e6e6e6'
             : '1px 0 0 #e6e6e6'
         }
-        href={props.href}
-        onClick={renderProps.onClick}
-        onFocus={renderProps.onFocus}
-        onBlur={renderProps.onBlur}
-        onMouseEnter={renderProps.onMouseEnter}
-        onMouseLeave={renderProps.onMouseLeave}
-        onMouseUp={renderProps.onMouseUp}
-        onMouseDown={renderProps.onMouseDown}
+        href={props.stub ? undefined : props.href}
+        onClick={props.stub ? undefined : renderProps.onClick}
+        onFocus={props.stub ? undefined : renderProps.onFocus}
+        onBlur={props.stub ? undefined : renderProps.onBlur}
+        onMouseEnter={props.stub ? undefined : renderProps.onMouseEnter}
+        onMouseLeave={props.stub ? undefined : renderProps.onMouseLeave}
+        onMouseUp={props.stub ? undefined : renderProps.onMouseUp}
+        onMouseDown={props.stub ? undefined : renderProps.onMouseDown}
       >
         <Flex
           align="center"
           justify="center"
           width={1}
           height={1}
-          children={props.children({
-            disabled: props.disabled || false,
-            hover: renderProps.hover,
-            focus: renderProps.focus,
-          })}
+          children={props.stub ? (
+            <Stub
+              height={6}
+              width={6}
+              r={12}
+            />
+          ) : (
+            props.children({
+              disabled: props.disabled || false,
+              hover: renderProps.hover,
+              focus: renderProps.focus,
+            })
+          )}
         />
       </CardLink>
     )}